+++ src/Csv/CsvProvider.fs
@@ -92,13 +92,19 @@ type public CsvProvider(cfg:TypeProviderConfig) as this =
         let body = csvErasedType?CreateEmpty () (Expr.Var rowToStringArrayVar, paramValue, replacer.ToRuntime headers,  sampleCsv.NumberOfColumns, separators, quote)
         Expr.Let(rowToStringArrayVar, rowToStringArray, body)))
       csvType.AddMember(ctor) 
-       
+
+      let parseRows = ProvidedMethod("ParseRows", [ProvidedParameter("text", typeof<string>)], rowType.MakeArrayType(), IsStaticMethod = true)
+      parseRows.InvokeCode <- fun (Singleton text) ->         
+        let body = csvErasedType?ParseRows () (text, Expr.Var stringArrayToRowVar, separators, quote, ignoreErrors)
+        Expr.Let(stringArrayToRowVar, stringArrayToRow, body)
+      csvType.AddMember parseRows
+
       { GeneratedType = csvType
         RepresentationType = csvType
         CreateFromTextReader = fun reader ->
           let body = 
             csvErasedType?Create () (Expr.Var stringArrayToRowVar, Expr.Var rowToStringArrayVar, replacer.ToRuntime reader, 
-                                              separators, quote, hasHeaders, ignoreErrors, skipRows, cacheRows)
+                                     separators, quote, hasHeaders, ignoreErrors, skipRows, cacheRows)
           Expr.Let(stringArrayToRowVar, stringArrayToRow, Expr.Let(rowToStringArrayVar, rowToStringArray, body))
         CreateFromTextReaderForSampleList = fun _ -> failwith "Not Applicable" }