+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@
 * Fix parsing of HTML closing tags with numbers.
 * Fixed handling of URI's with fragment but no query.
 * Fixed arrays created with XML provider having unneeded parent tags on some situations.
+* Allow to parse rows in CsvProvider without having to create a CsvFile.
 
 #### 2.1.1 - December 24 2014
 * Add SkipRows parameter to CsvProvider.